pyproject.toml
[project]
name = "uc-events"
version = "0.1.0"
description = ""
authors = [
    { name = "Alexander Spurlock", email = "alexanderspurlock@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.11.0, <4.0.0"
dependencies = [
    "aws-cdk-lib (>=2.223.0,<3.0.0)",
    "constructs (>=10.4.3,<11.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "feedparser (>=6.0.12,<7.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "boto3 (>=1.41.5,<2.0.0)",
    "boto3-stubs[full] (>=1.42.63,<2.0.0)",
]

[dependency-groups]
test = ["pytest (>=8.0)"]
dev = ["black (>=25.11.0,<26.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
filterwarnings = ["ignore::UserWarning"]
log_cli = true
log_cli_level = "INFO"

[tool.poetry]
package-mode = false
